Got Refactored Functional Comps To Work
